+++ R/operators.R
@@ -175,6 +175,14 @@
   torch_logical_not(x)
 }
 
+#' @rawNamespace if (getRversion() >= "4.3.0") S3method("%*%", torch_tensor)
+`%*%.torch_tensor` <- function(e1, e2) {
+  if (!is_torch_tensor(e2)) {
+    e2 <- torch_tensor(e2, device = e1$device)
+  }
+  torch_matmul(e1, e2)
+}
+
 #' @export
 dim.torch_tensor <- function(x) {
   cpp_tensor_dim(x$ptr)